change timestamp

+++ db/sqlc/account.sql.go
@@ -41,7 +41,7 @@ DELETE FROM accounts
 WHERE id = $1
 `
 
-func (q *Queries) DeleteAccount(ctx context.Context, id int32) error {
+func (q *Queries) DeleteAccount(ctx context.Context, id int64) error {
 	_, err := q.db.Exec(ctx, deleteAccount, id)
 	return err
 }
@@ -51,7 +51,7 @@ SELECT id, owner, balance, created_at FROM accounts
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetAccount(ctx context.Context, id int32) (Account, error) {
+func (q *Queries) GetAccount(ctx context.Context, id int64) (Account, error) {
 	row := q.db.QueryRow(ctx, getAccount, id)
 	var i Account
 	err := row.Scan(
@@ -108,7 +108,7 @@ RETURNING id, owner, balance, created_at
 `
 
 type UpdateAccountParams struct {
-	ID      int32 `json:"id"`
+	ID      int64 `json:"id"`
 	Balance int64 `json:"balance"`
 }
