+++ db/sqlc/entry.sql.go
@@ -20,7 +20,7 @@ RETURNING id, account_id, amount, created_at
 `
 
 type CreateEntryParams struct {
-	AccountID int32 `json:"account_id"`
+	AccountID int64 `json:"account_id"`
 	Amount    int64 `json:"amount"`
 }
 
@@ -41,7 +41,7 @@ SELECT id, account_id, amount, created_at FROM entries
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetEntry(ctx context.Context, id int32) (Entry, error) {
+func (q *Queries) GetEntry(ctx context.Context, id int64) (Entry, error) {
 	row := q.db.QueryRow(ctx, getEntry, id)
 	var i Entry
 	err := row.Scan(
@@ -62,7 +62,7 @@ OFFSET $3
 `
 
 type ListEntriesParams struct {
-	AccountID int32 `json:"account_id"`
+	AccountID int64 `json:"account_id"`
 	Limit     int32 `json:"limit"`
 	Offset    int32 `json:"offset"`
 }
