+++ db/sqlc/transfer.sql.go
@@ -20,8 +20,8 @@ INSERT INTO transfers (
 `
 
 type CreateTransferParams struct {
-	FromAccountID int32 `json:"from_account_id"`
-	ToAccountID   int32 `json:"to_account_id"`
+	FromAccountID int64 `json:"from_account_id"`
+	ToAccountID   int64 `json:"to_account_id"`
 	Amount        int64 `json:"amount"`
 }
 
@@ -43,7 +43,7 @@ SELECT id, from_account_id, to_account_id, amount, created_at FROM transfers
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetTransfer(ctx context.Context, id int32) (Transfer, error) {
+func (q *Queries) GetTransfer(ctx context.Context, id int64) (Transfer, error) {
 	row := q.db.QueryRow(ctx, getTransfer, id)
 	var i Transfer
 	err := row.Scan(
@@ -67,8 +67,8 @@ OFFSET $4
 `
 
 type ListTransfersParams struct {
-	FromAccountID int32 `json:"from_account_id"`
-	ToAccountID   int32 `json:"to_account_id"`
+	FromAccountID int64 `json:"from_account_id"`
+	ToAccountID   int64 `json:"to_account_id"`
 	Limit         int32 `json:"limit"`
 	Offset        int32 `json:"offset"`
 }
